% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantCommunitySim.R
\name{PlantCommunitySim}
\alias{PlantCommunitySim}
\title{Plant Community Assembly Simulation (PCAS)}
\usage{
PlantCommunitySim(
  L,
  jc_strength,
  sig_disp,
  tend,
  output,
  num_cores = 4,
  torus = TRUE
)
}
\arguments{
\item{L}{The grid size}

\item{jc_strength}{The strength of the Janzen-Connell (negative density dependence) effect. When 0, it represents a neutral model}

\item{sig_disp}{The dispersal width}

\item{tend}{The simulating time}

\item{output}{The directory and file name of the output}

\item{num_cores}{The number of the cores for parallel computing}

\item{torus}{With bundary or not}
}
\value{
None
}
\description{
Simulate a plant community assembly
}
